Studio Code: extract slash autocomplete into a useSlashCommands hook
Moves the inline slash-command popup, keyboard navigation, ARIA combobox
wiring, enter/exit animation, and the toolbar "/" toggle out of the Composer
into a dedicated useSlashCommands hook. The Composer now just spreads the
combobox props, defers keydown to the hook, renders the returned popup, and
wires the button to the toggle — keeping the component small. No behavior
change.
